backyard-produce/backend-produce-api/requirements.txt
# Backyard Produce API – FastAPI + PostgreSQL
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.3.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0,<5
python-multipart==0.0.18
